chore: add "typing :: typed" classifier, remove unused testing function
